requirements.txt
google-generativeai==0.3.0
streamlit==1.29.0
python-dotenv==1.0.0
langgraph==0.0.20
langchain==0.1.0
langchain-google-genai==0.0.6